Bulk update configuration
